Load client later
